+++ src/pvenv/subcommands/dvenv.py
@@ -12,8 +12,8 @@ if TYPE_CHECKING:
 class Command(BaseCommand):
     __slots__ = ()
 
-    def __init__(self, base_dir: Path, *, dry_run: bool, verbosity: int) -> None:
-        super().__init__(base_dir, dry_run=dry_run, verbosity=verbosity)
+    def __init__(self, base_dirs: list[Path], *, dry_run: bool, verbosity: int) -> None:
+        super().__init__(base_dirs, dry_run=dry_run, verbosity=verbosity)
 
     def run(self) -> None:
         self.execute(f"declare -f deactivate > {os.devnull} && deactivate")